Splitting up the tests to make it easier to test each component independantly
